+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod cli;
-pub mod config;
 pub mod get_limits;
 pub mod infra;
 pub mod notifications;